3aa96bd COMP: fixed some clang compiler warnings on the nightly dashboard

+++ Filtering/Testing/Cxx/TestTreeBFSIterator.cxx
@@ -54,7 +54,7 @@ int TestTreeBFSIterator(int, char *[])
     }
 
   //traverse the tree in a depth first fashion
-  for(int i = 0; i < correctSequence.size(); i++)
+  for(size_t i = 0; i < correctSequence.size(); i++)
     {
     if(!bfsIterator->HasNext())
       {
